+++ NeoAgi.AWS.CodeArtifact.Pruner/Models/EnumGeneric.cs
