+++ NeoAgi.AWS.CodeArtifact.Pruner/Models/PrunerConfig.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace NeoAgi.AWS.CodeArtifact.Pruner
+namespace NeoAgi.AWS.CodeArtifact.Pruner.Models
 {
     public class PrunerConfig
     {
