--- NeoAgi.AWS.CodeArtifact.Pruner/Package.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace NeoAgi.AWS.CodeArtifact.Pruner
-{
-    public class Package
-    {
-        public string Name {  get; set; } = string.Empty;
-        public string Namespace { get; set; } = string.Empty;
-        public string Format {  get; set; } = string.Empty;
-        public List<PackageVersion> Versions {  get; set; } = new List<PackageVersion>();
-    }
-}
