+++ NeoAgi.AWS.CodeArtifact.Pruner/Policies/PolicyManager.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using NeoAgi.AWS.CodeArtifact.Pruner.Models;
 
 namespace NeoAgi.AWS.CodeArtifact.Pruner.Policies
 {
